+++ app/src/main/res/menu/menu.xml
@@ -7,12 +7,4 @@
         android:icon="@drawable/ic_settings"
         android:title="@string/setting_menu"
         app:showAsAction="always" />
-
-    <item
-        android:id="@+id/delete"
-        android:icon="@drawable/ic_delete"
-        android:title="@string/delete_menu"
-        android:orderInCategory="4"
-        app:showAsAction="always" />
-
 </menu>
